+++ tui/terminal/wezterm/ethereal-gothic-angel.lua
@@ -1,0 +1,11 @@
+-- Ethereal-Theme (Gothic Angel)
+return {
+  foreground = "#eae3e7",
+  background = "#0f090b",
+  cursor_bg = "#ff3b70",
+  cursor_fg = "#0f090b",
+  selection_fg = "#eae3e7",
+  selection_bg = "#39121d",
+  ansi = { "#0f090b", "#ff3b70", "#ffffff", "#e8c4d0", "#e8c4d0", "#ff3b70", "#ff3b70", "#eae3e7" },
+  brights = { "#664e56", "#ff3b70", "#ffffff", "#ff7da0", "#ff5c86", "#ff3b70", "#ff3b70", "#ffffff" },
+}