+++ tui/terminal/wezterm/ethereal-luminous-space.lua
@@ -1,0 +1,11 @@
+-- Ethereal-Theme (Luminous Space)
+return {
+  foreground = "#e0ddf5",
+  background = "#0d0b1a",
+  cursor_bg = "#00ffd2",
+  cursor_fg = "#0d0b1a",
+  selection_fg = "#e0ddf5",
+  selection_bg = "#0b363a",
+  ansi = { "#0d0b1a", "#ff9de0", "#ffffff", "#bfdbfe", "#00ffd2", "#00ffd2", "#00ffd2", "#e0ddf5" },
+  brights = { "#6c698f", "#ff9de0", "#ffffff", "#ff9de0", "#3cffdc", "#cf9dff", "#00ffd2", "#ffffff" },
+}